+++ client/shared/chat-card/index.tsx
@@ -65,6 +65,7 @@ export const ChatCard = forwardRef<HTMLDivElement, ChatCardProps>(
                 h5: (props) => (
                   <h5 {...props} className="dark:text-slate-300" />
                 ),
+                a: (props) => <a {...props} className="dark:text-slate-400" />,
               }}
             >
               {message.text}