+++ server/database/interview.py
@@ -1,5 +1,7 @@
 from enum import Enum
 from typing import List, Optional
+
+from beanie import Indexed
 from database.base import BaseMongoModel
 
 
@@ -11,7 +13,7 @@ class InterviewType(str, Enum):
 class Interview(BaseMongoModel):
     name: str
     description: str
-    url: Optional[str] = None
+    url: Indexed(str, unique=True) = None  # type:ignore
     organization_name: Optional[str] = None
     interview_type: InterviewType = InterviewType.GENERAL
     position: Optional[str] = None